# docs/openapi.yaml
openapi: 3.0.3
info:
  title: iDME 工业 MiniApp API
  description: |
    华为 iDME 数据建模引擎 - 工业 MiniApp 接口文档。

    ## 快速开始

    ### 推荐创建顺序（依赖链）

    **Level 0（无依赖）**：Unit（计量单位）, BusinessPartner（往来单位）, EquipmentClassfication（设备分类）, PartClassfication（物料分类）

    **Level 1**：Department（部门）, Employee（员工）

    **Level 2**：EquipmentModel（设备型号）, Location（位置）

    **Level 3**：Equipment（设备）, Part（物料）

    **Level 4**：WorkingProcedure（工序）, WorkingPlan（工艺路线）

    **Level 5**：BOMItem（BOM项）, PlanProcedureLink（计划-工序关联）, ProcedureEquipmentLink（工序-设备关联）, ProcedurePartLink（工序-物料关联）, EquipmentSparePartLink（设备-备件关联）

    ### 注意事项
    - **引用字段**（如 `manufacturerName`, `category`）需要传入对应模块的 **ID**，不是文本名称
    - **Part** 和 **WorkingPlan** 是版本对象，`WorkingPlan.productPart` 引用 Part 时使用 `master.id` 而非实例 `id`
    - **枚举值**使用英文（如 `Standby`, `Active`, `Machinery`）
    - 所有接口的 example 字段已填充真实可用的示例ID，可直接复制使用

    ### 示例数据
    - 固定ID文档: `/docs/openapi-examples-ids.md`
    - 测试文件: `/src/test/java/com/tsukilc/idme/integration/OpenApiExamplesDataTest.java`

    ---

    - 管理界面: http://99.suyiiyii.top:8003/rdm_a1b52ff379ee46ed8928d7f2ceb908f6_app/services/index.html
    - 健康检查: http://99.suyiiyii.top:8003/rdm_a1b52ff379ee46ed8928d7f2ceb908f6_app/services/v1/health

    认证: Basic Auth 用户名 sysadmin / 密码见项目说明。
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://99.suyiiyii.top:8003/rdm_a1b52ff379ee46ed8928d7f2ceb908f6_app/services
    description: iDME SDK 服务
security:
  - basicAuth: []
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: 用户名 sysadmin，密码见项目文档
  parameters:
    pageNum:
      name: pageNum
      in: query
      description: 页码，从 1 开始
      schema:
        type: integer
        default: 1
    pageSize:
      name: pageSize
      in: query
      description: 每页大小
      schema:
        type: integer
        default: 20
    idPath:
      name: id
      in: path
      required: true
      description: 主键 ID
      schema:
        type: string
  schemas:
    ApiResponse:
      type: object
      properties:
        result:
          type: string
          example: SUCCESS
        data: {}
        errors:
          type: array
          items:
            type: string
      description: 统一响应包装
...
    WorkingProcedure:
      type: object
      properties:
        id: { type: string }
        procedureCode: { type: string }
        procedureName: { type: string }
        steps: { type: string }
        mainProductionEquipment: { type: string }
        mainInspectionEquipment: { type: string }
        operatorUser: { type: string }
        operatorRef: { type: string }
        status: { type: string, enum: [未开始, 进行中, 已完成, 暂停] }
        remarks: { type: string }
    WorkingProcedureCreate:
      type: object
      required: [procedureCode, procedureName]
      properties:
        procedureCode:
          type: string
          description: 工序编码
          example: "WP-DEMO"
        procedureName:
          type: string
          description: 工序名称
          example: "粗加工"
        steps:
          type: string
          description: 工序步骤
          example: "1.上料 2.加工 3.检测"
        mainProductionEquipment:
          type: string
          description: 主生产设备ID（-> Equipment模块）
          example: "6000000001"
        mainInspectionEquipment:
          type: string
          description: 主检测设备ID（-> Equipment模块，SDK必填）
          example: "6000000002"
        operatorUser:
          type: string
          description: 操作员用户ID（-> User模块，可选）
        operatorRef:
          type: string
          description: 操作员ID（-> Employee模块）
          example: "4000000002"
        startTime:
          type: string
          format: date
          description: 开始时间（日期格式，如2025-04-05）
        endTime:
          type: string
          format: date
          description: 结束时间（日期格式，如2025-04-10）
        status:
          type: string
          enum: [NotStarted, InProgress, Completed, Paused]
          description: "工序状态：未开始/进行中/已完成/已暂停"
          example: "NotStarted"
        remarks:
          type: string
          description: 备注
...
    ProcedureStatusUpdate:
      type: object
      properties:
        status: { type: string }
        actualStart: { type: string, format: date-time }
        actualEnd: { type: string, format: date-time }
...
    WorkingProcedureBatch:
      type: object
      properties:
        procedures:
          type: array
          items:
            type: object
            properties:
              procedureCode: { type: string }
              procedureName: { type: string }
...
paths:
...
  /api/working-procedure:
    get:
      tags: [工序管理]
      summary: 分页查询工序列表
      operationId: listWorkingProcedure
      parameters:
        - $ref: '#/components/parameters/pageNum'
        - $ref: '#/components/parameters/pageSize'
        - name: procedureName
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          total: { type: integer }
                          pageNum: { type: integer }
                          pageSize: { type: integer }
                          records:
                            type: array
                            items: { $ref: '#/components/schemas/WorkingProcedure' }
    post:
      tags: [工序管理]
      summary: 创建工序
      operationId: createWorkingProcedure
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkingProcedureCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkingProcedure'
  /api/working-procedure/batch:
    post:
      tags: [工序管理]
      summary: 批量创建工序
      operationId: batchCreateWorkingProcedure
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkingProcedureBatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/working-procedure/{id}:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    get:
      tags: [工序管理]
      summary: 查询工序详情
      operationId: getWorkingProcedure
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkingProcedure'
    put:
      tags: [工序管理]
      summary: 更新工序
      operationId: updateWorkingProcedure
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkingProcedureCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      tags: [工序管理]
      summary: 删除工序
      operationId: deleteWorkingProcedure
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/working-procedure/{id}/status:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    patch:
      tags: [工序管理]
      summary: 更新工序状态
      operationId: updateWorkingProcedureStatus
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProcedureStatusUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...